@@ -4,7 +4,8 @@
 // permitted by law, all liability for your use of the code is disclaimed. This source code is governed by Apache
 // License 2.0 that can be found in the LICENSE file.
 
-//+build arm arm64,!darwin
+//go:build arm || (arm64 && !darwin)
+// +build arm arm64,!darwin
 
 package log
 
