@@ -56,6 +56,8 @@ func TestAccount(t *testing.T) {
 	addr2 := account2.PublicKey().Address()
 	r.NotNil(addr2)
 	r.False(IsSignerExist(addr2.String()))
+	_, err = keyStoreAccountToPrivateKey(addr2.String(), passwd)
+	r.Contains(err.Error(), "does not match all local keys")
 	filePath := sm2KeyPath(addr2)
 	addrString, err := storeKey(account2.HexString(), config.ReadConfig.Wallet, passwd)
 	r.NoError(err)
@@ -72,7 +74,7 @@ func TestAccount(t *testing.T) {
 
 	// test keystore conversion and signing
 	CryptoSm2 = false
-	prvKey, err := LocalAccountToPrivateKey(addr.String(), passwd)
+	prvKey, err := keyStoreAccountToPrivateKey(addr.String(), passwd)
 	r.NoError(err)
 	msg := hash.Hash256b([]byte(nonce))
 	sig, err := prvKey.Sign(msg[:])
@@ -80,7 +82,7 @@ func TestAccount(t *testing.T) {
 	r.True(prvKey.PublicKey().Verify(msg[:], sig))
 
 	CryptoSm2 = true
-	prvKey2, err := LocalAccountToPrivateKey(addr2.String(), passwd)
+	prvKey2, err := keyStoreAccountToPrivateKey(addr2.String(), passwd)
 	r.NoError(err)
 	msg2 := hash.Hash256b([]byte(nonce))
 	sig2, err := prvKey2.Sign(msg2[:])
