@@ -30,6 +30,7 @@ import (
 
 	"github.com/iotexproject/iotex-core/pkg/log"
 	"github.com/iotexproject/iotex-core/pkg/routine"
+	"github.com/iotexproject/iotex-core/pkg/tracer"
 )
 
 const (
@@ -312,7 +313,10 @@ func (p *Agent) Stop(ctx context.Context) error {
 }
 
 // BroadcastOutbound sends a broadcast message to the whole network
-func (p *Agent) BroadcastOutbound(_ context.Context, msg proto.Message) (err error) {
+func (p *Agent) BroadcastOutbound(ctx context.Context, msg proto.Message) (err error) {
+	_, span := tracer.NewSpan(ctx, "Agent.BroadcastOutbound")
+	defer span.End()
+
 	host := p.host
 	if host == nil {
 		return ErrAgentNotStarted
